preparazione-esame/django/my_blog_api/pyproject.toml
[tool.poetry]
name = "my-blog-api"
version = "0.1.0"
description = ""
authors = ["ilveron <alessandromonetti@outlook.it>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.1"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.3.1"
coreapi = "^2.3.3"
pyyaml = "^6.0.1"
dj-rest-auth = "^5.0.2"
django-allauth = "^0.55.0"
pytest-django = "^4.7.0"
mixer = "^7.2.2"
pytest = "^7.4.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
